attemps fixing cocos2d/cocos2d-x/19188 (WARNING: see note below!)
Attempts to fix a disappearing Ended/Cancel event and a memory reuse.

It might introduce a small memory leak as touches are retained in the
Began event, but never released as the Ended/Cancel never arrives.

The proper solution should be investigate why the touches are reused
but the event is never called leading to this bug.

See https://github.com/cocos2d/cocos2d-x/issues/19188 for full details.